# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gin快速启动脚手架",
        "title": "gin-fast API",
        "termsOfService": "https://tianbaoo.github.io/",
        "contact": {
            "name": "Tianbao",
            "url": "https://tianbaoo.github.io/",
            "email": "gtb365@163.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:7890",
    "basePath": "/api",
    "paths": {
        "/api/login": {
            "post": {
                "description": "用户登录接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户登录接口",
                "parameters": [
                    {
                        "description": "请求body中需携带的参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializers.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"返回成功\",\"data\":{\"email\":\"gtb365@163.com\",\"id\":93997900482051,\"name\":\"\",\"phone\":\"15926187015\",\"token\":\"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2NDAwNjczMzksImlkIjo5Mzk5NzkwMDQ4MjA1MSwidXNlcm5hbWUiOiJ0aWFuYmFvIn0.luFD6rMu3eYrnvWaIu8uBUiKszFGnkFrf8b9-0djnPc\",\"username\":\"tianbao\"}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"code\": 400, \"msg\": \"返回错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "description": "用户注册接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户注册接口",
                "parameters": [
                    {
                        "description": "请求body中需携带的参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializers.Register"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"msg\": \"返回成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"code\": 400, \"msg\": \"返回错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "serializers.Login": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "serializers.Register": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}